Misc. changes to support MSVC builds

+++ src/display/class_pointer.cpp
@@ -40,7 +40,7 @@ static FunctionField mthGrabX11Pointer[] = { { "Surface", FD_LONG }, { NULL, 0 }
 #endif
 
 static LONG glDefaultSpeed = 160;
-static FLOAT glDefaultAcceleration = 0.8;
+static DOUBLE glDefaultAcceleration = 0.8;
 static TIMER glRepeatTimer = 0;
 
 static ERROR repeat_timer(extPointer *, LARGE, LARGE);
